+++ examples/mundialis_extended.py
@@ -80,8 +80,7 @@ def execute(
         image_collection,
         json.dumps(connection.describe_collection(image_collection), indent=2),
     )
-    # TODO: change to con.load_collection()
-    cube = ImageCollectionClient.load_collection(session = connection, collection_id = image_collection, bands = all_bands)
+    cube = connection.load_collection(collection_id=image_collection, bands=all_bands)
     cube = cube.filter_bbox( **bbox)
     cube = cube.filter_temporal(extent=temporal_extent)