+++ examples/mundialis_mini.py
@@ -20,11 +20,9 @@ temporal_extent=["2018-05-01T00:00:00.000Z","2018-10-01T00:00:00.000Z"]
 spectral_extent = ["S2_8", "S2_4", "S2_2"]
 
 # connect to mundialis backend
-session = openeo.connect(backend_url).authenticate_basic()
+connection = openeo.connect(backend_url).authenticate_basic()
 
-# TODO change to s2_radiometry = session.load_collection( ...)
-s2_radiometry = ImageCollectionClient.load_collection(
-                    session=session,
+s2_radiometry = connection.load_collection(
                     collection_id=collection_id,
                     temporal_extent=temporal_extent,
                     spatial_extent=spatial_extent