+++ tests/test_rest_session.py
@@ -56,6 +56,7 @@ PROCESSES = [
 
 @requests_mock.mock()
 class TestUserFiles(TestCase):
+    # TODO: review this test class (looks very dated), most things are probably already covered elsewhere
 
     def setUp(self):
         # configuration phase: define username, endpoint, parameters?
@@ -121,7 +122,7 @@ class TestUserFiles(TestCase):
 
     def test_list_capabilities(self, m):
         capabilities = {
-            "api_version": "0.4.0",
+            "api_version": "1.0.0",
             "endpoints": [
                 {"path": "/collections", "methods": ["GET"]},
             ]
@@ -142,16 +143,18 @@ class TestUserFiles(TestCase):
             openeo.connect(self.endpoint)
 
     def test_capabilities_api_version_recent(self, m):
-        m.register_uri('GET', "{}/".format(self.endpoint), json={'version': '0.4.0'})
+        m.register_uri("GET", "{}/".format(self.endpoint), json={"version": "1.0.0"})
         capabilities = openeo.connect(self.endpoint).capabilities()
-        assert capabilities.version() == '0.4.0'
-        assert capabilities.api_version() == '0.4.0'
+        assert capabilities.version() == "1.0.0"
+        assert capabilities.api_version() == "1.0.0"
 
     def test_capabilities_api_version_recent2(self, m):
-        m.register_uri('GET', "{}/".format(self.endpoint), json={'api_version': '0.4.1'})
+        m.register_uri(
+            "GET", "{}/".format(self.endpoint), json={"api_version": "1.1.0"}
+        )
         capabilities = openeo.connect(self.endpoint).capabilities()
-        assert capabilities.version() == '0.4.1'
-        assert capabilities.api_version() == '0.4.1'
+        assert capabilities.version() == "1.1.0"
+        assert capabilities.api_version() == "1.1.0"
 
     def test_capabilities_api_version_check(self, m):
         capabilties_url = "{}/".format(self.endpoint)
@@ -168,7 +171,7 @@ class TestUserFiles(TestCase):
         assert capabilities.api_version_check.above('1.2.2')
 
     def test_list_collections(self, m):
-        m.get("http://localhost:8000/api/", json={"api_version": "0.4.0"})
+        m.get("http://localhost:8000/api/", json={"api_version": "1.0.0"})
         con = openeo.connect(self.endpoint)
 
         collection_url = "{}/collections".format(self.endpoint)
@@ -177,7 +180,7 @@ class TestUserFiles(TestCase):
         assert collections == COLLECTIONS
 
     def test_get_collection(self, m):
-        m.get("http://localhost:8000/api/", json={"api_version": "0.4.0"})
+        m.get("http://localhost:8000/api/", json={"api_version": "1.0.0"})
         con = openeo.connect(self.endpoint)
 
         collection_org = COLLECTIONS[0]
@@ -188,7 +191,7 @@ class TestUserFiles(TestCase):
         assert collection == collection_org
 
     def test_get_all_processes(self, m):
-        m.get("http://localhost:8000/api/", json={"api_version": "0.4.0"})
+        m.get("http://localhost:8000/api/", json={"api_version": "1.0.0"})
         con = openeo.connect(self.endpoint)
 
         processes_url = "{}/processes".format(self.endpoint)